add player characters AI (wip)

+++ Assets/Code/DungeonTeam/CharacterSkill/Skills/FireballSkill/Base/FireballSkillModelBase.cs
@@ -8,7 +8,7 @@ public abstract class FireballSkillModelBase : Model
 	public abstract string SkillName { get; }
 	public abstract float FireballSpeed { get; }
 	public abstract bool IsReadyToActivate { get; }
-	public abstract float FireballDamage { get; }
+	public abstract int FireballDamage { get; }
 
     public abstract void ActivateSkill();
 	public abstract void UpdateFireballData(FireballData fireballData);
